adding options for animated gif capturing (Mac only via wacaw)
Using wacaw and ffmpeg (and convert) it is possible to generate and annotate a
short animated gif for your lolcommit.  Use -a 3 or --animate=3 option in your
.git/hooks/post-commit call to capture and animate (for 3 second gif) The longer
the time, the larger the animated gif file-size. Some extra points to note;

* currently mac only
* if -a 0 (or not present) lolcommits will operate as normal
* defaults to a 320x240 sized animated gif
* wacaw is open source here; http://webcam-tools.sourceforge.net

I've tested with a few of the plugins and it appears to work OK! Also updated
README.md with instructions on how to enable and configure