Adding a better feature for testing the --fork option
Adding the concept of an lolcommits.pid file to hold the child process id when
running with the --fork option (introduced in this previous PR
`https://github.com/mroth/lolcommits/pull/109`)

With a pid file, it is possible to test that the child process has exited in
`lolcommits.feature`. A new step waits for the file to be removed. To allow
command output step to work, `$stdout.sync = true` is set before forking.