remove active support from gemspec
all usage of it seems to have been confined to what I just refactored
out of runner.

TODO: tests are passing but still need to audit the existing plugins,
as some of them may have taken advantage of the fact we had it loaded,
and since the plugins general have poor test coverage its possible this
breaks one.