switch to container-based docker builds
travis-ci now is testing docker container-based builds, which have a lot
of advantages and could speed things up significantly:

    http://blog.travis-ci.com/2014-12-17-faster-builds-with-container-based-infrastructure/

this mean no sudo, so we cant use apt-get to install mplayer and ffmpeg.
just commenting them out for now which should cause a failure, but that
will be the starting point to working around them.

also enabling bundler cache, which is available for free projects with
this build system, and should speed up installs greatly.