use default case to get rid of magic variables
many of the scenarios used a magic string for the directory name, which
then had to be passed on to other steps. this commit introduces a
default case for repo name, and handles the corresponding loldir with
convenience functions so a made up directory name is not needed in most
cases.