Avoid invoking ruby if in a rebase
Invoking ruby and loading lolcommits can be very slow. For example, on
my (fairly fast) machine:

$ time ruby -r lolcommits -e ''
ruby -r lolcommits -e ''  0.61s user 0.10s system 99% cpu 0.716 total

This isn't usually an issue for a single commit, but it makes lolcommits
impossible to use with a rebase-driven workflow.

This adds an if guard to the post-commit hook installer, which entirely
skips running ruby when in the middle of a rebase.