make test and debug mode global flags, init config with test everywhere
Clamp flags are only inherited with anonymous subclass block syntax for
subcommands, so instead need to put them in our wrapper class (knew it
would come in handy!).  Some CLI managers have a separate concept for a
global flag that exists between invocation ARGV[0] and the subcommand
syntax wise, but Clamp does not, so keep things simple.

In addition, need to make sure we pass the test mode flag on everywhere
we instantiate the Configuration class.  This is suggesting we can
handle this instantiation once globally somewhere, but need to think
about refactoring that.